Refactor - move functions from util/summer_term.py into sws/section.py.
